refactor useSpottable into separate file
